Changing types of monetary values ​​from Double to BigDecimal and improving the logic of the createBudget method

+++ src/main/java/com/bandeira/clinica_alves_oliveira/dtos/UpdatePaymentDTO.java
@@ -2,11 +2,12 @@ package com.bandeira.clinica_alves_oliveira.dtos;
 
 import com.bandeira.clinica_alves_oliveira.models.FormOfPayment;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record UpdatePaymentDTO(
 
-         Double value,
+         BigDecimal value,
 
          String namePatient,
