+++ src/main/java/com/bandeira/clinica_alves_oliveira/dtos/UpdateProceduresDTO.java
@@ -2,11 +2,13 @@ package com.bandeira.clinica_alves_oliveira.dtos;
 
 import com.bandeira.clinica_alves_oliveira.models.Specialty;
 
+import java.math.BigDecimal;
+
 public record UpdateProceduresDTO(
 
         String description,
 
         Specialty specialty,
 
-        Double value) {
+        BigDecimal value) {
 }
